quantlib_wrapper into utils
